+++ backend/src/main/java/org/bihealth/mi/risk_assessment_api/repository/dataset/DatasetTableAttributeRepository.java
@@ -4,5 +4,5 @@ import org.bihealth.mi.risk_assessment_api.model.dataset.DatasetTableAttribute;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface DatasetTableAttributeRepository extends JpaRepository<DatasetTableAttribute, Integer> {
+public interface DatasetTableAttributeRepository extends JpaRepository<DatasetTableAttribute, Long> {
 }
